keep linux files
linux build also removes files that arent needed

+++ src/scripts/removeUnneededFiles.js
@@ -49,9 +49,10 @@ exports.default = async function(context) {
         files.forEach(file => {
             // List of files to keep in the main directory
             const filesToKeep = ['App Launcher.exe', 'ffmpeg.dll', 'icudtl.dat', 'resources.pak', 'v8_context_snapshot.bin', 'resources', 'locales', 'chrome_100_percent.pak', 'chrome_200_percent.pak'];
+            const filesToKeepLinux = ['app-launcher', 'chrome_crashpad_handler', 'chrome-sandbox', 'libffmpeg.so', 'icudtl.dat', 'resources.pak', 'v8_context_snapshot.bin', 'resources', 'locales', 'chrome_100_percent.pak', 'chrome_200_percent.pak'];
             
             // If the file is not in the list, delete it
-            if (!filesToKeep.includes(file)) {
+            if (!filesToKeep.includes(file) && !filesToKeepLinux.includes(file)) {
                 const filePath = path.join(mainDir, file);  // <-- Fixed this to use mainDir, not localeDir
                 try {
                     fs.unlinkSync(filePath);
